requirements.txt
streamlit==1.29.0
python-dotenv==1.0.0
google-generativeai
langchain-google-genai